// src/types.ts
export interface CompanyData {
  name: string
  unified_code: string
  legal_person: string
  registered_capital: string
  founded_date: string
  status: string
  address: string
  scope: string
  industry: string
  reg_auth: string
  reg_number: string
  org_code: string
  tax_id: string
  employees: string
  contacts: string[]
  investors: string[]
  branches: string[]
  source: string
}